csv_reader now returns 2d tensor of 10dim vectors
Updated csv reader so it now returns sample data of palettes with separate rgb channel values. Each palette is now represented as a 9 element vector where each consecutive triplet represents the normalized rgb values of one color in the palette. The label data (the ratings of the palettes) is changed from being 1, 2, 3 and insted now is represented as 0, 1, 2.